feat: refactor user avatar components and update imports
