+++ src/components/user/DefaultUserAvatar.tsx
@@ -1,0 +1,5 @@
+import { UserIcon } from '@heroicons/react/24/solid'
+
+export const DefaultUserAvatar = () => {
+  return <UserIcon className='size-[64px] rounded-full bg-gray-400 p-1 text-white' />
+}
