+++ src/components/user/avatar.tsx
@@ -1,9 +1,5 @@
 import Image from 'next/image'
 
-type Props = {
-  imageUrl: string
-}
-
-export const Avatar = ({ imageUrl }: Props) => {
+export const Avatar = ({ imageUrl }: { imageUrl: string }) => {
   return <Image src={imageUrl} alt='user image' width={64} height={64} className='rounded-full' />
 }
